fix clientServiceImpl et autres broutilles client

+++ service-impl/src/main/java/fr/uga/l3miage/photonum/service/ImageServiceImpl.java
@@ -1,9 +1,12 @@
 package fr.uga.l3miage.photonum.service;
 import fr.uga.l3miage.photonum.data.domain.Image;
+import fr.uga.l3miage.photonum.data.repo.ImageRepository;
+import fr.uga.l3miage.photonum.data.domain.Photo;
 import fr.uga.l3miage.photonum.service.base.BaseService;
 import fr.uga.l3miage.photonum.service.EntityNotFoundException;
 
 import java.util.Collection;
+import java.util.Set;
 
 import org.springframework.beans.factory.annotation.Autowired;
 
@@ -38,7 +41,10 @@ public class ImageServiceImpl implements ImageService{
         if (im == null) {
             throw new EntityNotFoundException("l'image avec id=%d n'a pas été trouvée".formatted(id));
         }
-
+        Set<Photo> photos = im.getPhotos();
+        if (photos.size() > 0){
+            throw new Exception("Il existe des photos de cette image, elle ne peut pas être supprimée");
+        }
         imageRepository.delete(im);
 
     }
